moving more responsibilities into LinqQuery
