+++ src/Marten/Linq/QueryModelExtensions.cs
@@ -2,16 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using Baseline;
-using Baseline.Conversion;
-using Marten.Events;
-using Marten.Linq.Model;
 using Marten.Schema;
-using Marten.Transforms;
-using Marten.Util;
 using Remotion.Linq;
 using Remotion.Linq.Clauses;
 using Remotion.Linq.Clauses.Expressions;
-using Remotion.Linq.Clauses.ResultOperators;
 
 namespace Marten.Linq
 {
@@ -83,64 +77,6 @@ namespace Marten.Linq
                        AdditionalFromClause;
         }
 
-        public static ISelector<T> BuildSelector<T>(this IDocumentSchema schema, IQueryableDocument mapping,
-            QueryModel query)
-        {
-            var selectable = query.AllResultOperators().OfType<ISelectableOperator>().FirstOrDefault();
-            if (selectable != null)
-            {
-                return selectable.BuildSelector<T>(schema, mapping);
-            }
-
-
-            if (query.HasSelectMany())
-            {
-                return buildSelectorForSelectMany<T>(mapping, query, schema.StoreOptions.Serializer());
-            }
-
-            if (query.SelectClause.Selector.Type == query.SourceType())
-            {
-                if (typeof(T) == typeof(string))
-                    return (ISelector<T>) new JsonSelector();
-
-                // I'm so ashamed of this hack, but "simplest thing that works"
-                if (typeof(T) == typeof(IEvent))
-                    return mapping.As<EventQueryMapping>().Selector.As<ISelector<T>>();
-
-                var resolver = schema.ResolverFor<T>();
-                return new WholeDocumentSelector<T>(mapping, resolver);
-            }
-
-
-
-            return createSelectTransformSelector<T>(schema, mapping, query);
-        }
-
-        [Obsolete("Going to replace this with the new LinqQuery model soon")]
-        public static SelectManyQuery ToSelectManyQuery(this QueryModel query, IQueryableDocument mapping)
-        {
-            return new SelectManyQuery(mapping, query, 0);
-        }
-
-        private static ISelector<T> buildSelectorForSelectMany<T>(IQueryableDocument mapping, QueryModel query, ISerializer serializer)
-        {
-            return query.ToSelectManyQuery(mapping).ToSelector<T>(serializer);
-        }
-
-        private static ISelector<T> createSelectTransformSelector<T>(IDocumentSchema schema, IQueryableDocument mapping,
-            QueryModel query)
-        {
-            var visitor = new SelectorParser(query);
-            visitor.Visit(query.SelectClause.Selector);
-
-            return visitor.ToSelector<T>(schema, mapping);
-        }
-
-        public static ISelector<T> BuildSelector<T>(this IDocumentSchema schema, QueryModel query)
-        {
-            var mapping = schema.MappingFor(query.SourceType()).ToQueryableDocument();
-            return schema.BuildSelector<T>(mapping, query);
-        }
 
         public static IDocumentMapping MappingFor(this IDocumentSchema schema, QueryModel model)
         {
